+++ lib/providers/presets.dart
@@ -34,6 +34,7 @@ class PresetProvider {
   late final LocationMatcher locationMatcher;
   bool ready = false;
   final Ref _ref;
+  final Map<String, PresetField> _fieldCache = {};
 
   PresetProvider(this._ref) {
     initMatcher();
@@ -352,8 +353,17 @@ class PresetProvider {
       if (seenFields.contains(row['name'])) continue;
       if (row['name'] == 'opening_hours/covid19') continue;
       seenFields.add(row['name'] as String);
-      final options = await _getComboOptions(row);
-      final field = fieldFromJson(row, options: options);
+
+      // Either build a field, or restore it from a cache.
+      PresetField field;
+      if (_fieldCache.containsKey(row['name'])) {
+        field = _fieldCache[row['name']]!;
+      } else {
+        final options = await _getComboOptions(row);
+        field = fieldFromJson(row, options: options);
+        _fieldCache[row['name'] as String] = field;
+      }
+
       // query options if needed
       if (row['required'] == 1) {
         fields.add(field);
@@ -381,6 +391,12 @@ class PresetProvider {
 
   Future<Map<String, PresetField>> _getFields(
       List<String> names, Locale? locale) async {
+    // There's a chance all the fields were cached.
+    final nonCachedNames =
+        names.where((element) => !_fieldCache.containsKey(element)).toList();
+    if (nonCachedNames.isEmpty)
+      return {for (final name in names) name: _fieldCache[name]!};
+
     if (!ready) await _waitUntilReady();
     final langCTE = _localeCTE(locale);
     final params = List.filled(names.length, '?').join(',');
@@ -398,13 +414,23 @@ class PresetProvider {
     order by lscore
     ''';
     final results = await _db!.rawQuery(sql, names);
+
     Map<String, PresetField> fields = {};
     final seenFields = <String>{};
     for (final row in results) {
       if (seenFields.contains(row['name'])) continue;
       seenFields.add(row['name'] as String);
-      final options = await _getComboOptions(row);
-      final field = fieldFromJson(row, options: options);
+
+      // Either build a field, or restore it from a cache.
+      PresetField field;
+      if (_fieldCache.containsKey(row['name'])) {
+        field = _fieldCache[row['name']]!;
+      } else {
+        final options = await _getComboOptions(row);
+        field = fieldFromJson(row, options: options);
+        _fieldCache[row['name'] as String] = field;
+      }
+
       fields[row['name'] as String] = field;
     }
     return fields;